requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pyyaml>=6.0
tqdm>=4.62.0
requests>=2.26.0
pandas>=1.3.0

# Vector database
faiss-cpu>=1.7.2
# For GPU support: faiss-gpu>=1.7.2

# Embedding models
sentence-transformers>=2.2.0
transformers>=4.18.0
torch>=1.10.0

# Data processing
nltk>=3.6.5
scikit-learn>=1.0.0

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.12.0

# API (optional)
fastapi>=0.78.0
uvicorn>=0.17.0
pydantic>=1.9.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0